+++ lib/features/contacts/participants_selection_provider.dart
@@ -3,7 +3,7 @@ import 'dart:collection';
 import 'package:ejyption_time_2/core/Contacts/contacts_impl_flutter_contacts.dart';
 import 'package:ejyption_time_2/core/Contacts/contacts_provider_interface.dart';
 import 'package:ejyption_time_2/core/global_model.dart';
-import 'package:ejyption_time_2/models/meeting.dart';
+import 'package:ejyption_time_2/models/meeting/meeting.dart';
 import 'package:ejyption_time_2/models/perticipants.dart';
 import 'package:flutter/widgets.dart';
