+++ lib/features/contacts/participants_selection_widget.dart
@@ -4,7 +4,7 @@ import 'package:ejyption_time_2/core/common_widgets/avatar.dart';
 import 'package:ejyption_time_2/core/common_widgets/main_popup_menu.dart';
 import 'package:ejyption_time_2/core/common_widgets/templates.dart';
 import 'package:ejyption_time_2/features/contacts/participants_selection_provider.dart';
-import 'package:ejyption_time_2/models/meeting.dart';
+import 'package:ejyption_time_2/models/meeting/meeting.dart';
 import 'package:ejyption_time_2/models/my_contact.dart';
 import 'package:ejyption_time_2/models/perticipants.dart';
 import 'package:flutter/material.dart';
