+++ lib/features/detailed_meeting/constant_field_widget.dart
@@ -6,8 +6,8 @@ import 'package:ejyption_time_2/core/main_functions.dart';
 import 'package:ejyption_time_2/features/detailed_meeting/constant_field_provider.dart';
 import 'package:ejyption_time_2/features/modify_meeting/field_edit_form.dart';
 import 'package:ejyption_time_2/features/modify_meeting/modifying_field_provider.dart';
-import 'package:ejyption_time_2/models/meeting.dart';
-import 'package:ejyption_time_2/models/negotiating_field.dart';
+import 'package:ejyption_time_2/models/meeting/meeting.dart';
+import 'package:ejyption_time_2/models/negotiating_fields/negotiating_field.dart';
 import 'package:ejyption_time_2/models/point_assestment.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
