+++ lib/features/detailed_meeting/meeting_detailed_provider.dart
@@ -5,7 +5,7 @@ import 'package:ejyption_time_2/features/detailed_meeting/constant_field_provide
 import 'package:ejyption_time_2/models/my_contact.dart';
 import 'package:ejyption_time_2/models/participant.dart';
 import 'package:ejyption_time_2/models/probability_assesstment.dart';
-import '../../models/meeting.dart';
+import '../../models/meeting/meeting.dart';
 import 'package:flutter/widgets.dart';
 
 enum MainMenu {
