+++ lib/models/participant.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 
 import 'package:ejyption_time_2/core/global_model.dart';
-import 'package:ejyption_time_2/models/meeting.dart';
+import 'package:ejyption_time_2/models/meeting/meeting.dart';
 import 'package:ejyption_time_2/models/my_contact.dart';
 import 'package:flutter/material.dart';
